LayoutTests/http/tests/security/drag-drop-different-origin.html

<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

async function moveToCenter(element)
{
    x = element.offsetParent.offsetLeft + element.offsetLeft + element.offsetWidth / 2;
    y = element.offsetParent.offsetTop + element.offsetTop + element.offsetHeight / 2;
    await eventSender.asyncMouseMoveTo(x, y);
}

async function runTest() {

    var x, y;
    var span = document.getElementById("span");
    moveToCenter(span);
    await eventSender.asyncMouseDown();
    await eventSender.asyncMouseUp();
    await eventSender.asyncMouseDown();
    await eventSender.asyncMouseUp();

    eventSender.leapForward(1000);

    await eventSender.asyncMouseDown();

    eventSender.leapForward(500);

    var input = document.getElementById("target");
    moveToCenter(input);
    eventSender.leapForward(500);
    await eventSender.asyncMouseUp();
    
    input.contentWindow.postMessage("go", "*");
}
</script>
</head>
<body onload="runTest()">
<span id="span">Dragme</span>
<iframe id="target" src="http://localhost:8000/security/resources/drag-drop.html"></iframe>
</body>
</html>
