third_party/blink/web_tests/http/tests/security/drag-drop-different-origin-expected.txt

CONSOLE MESSAGE: line 9: PASS
Dragme 
